Skip marks popup after bosses
